% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_enedis.R
\name{access_enedis}
\alias{access_enedis}
\title{access_enedis}
\usage{
access_enedis(url, realm = "particuliers", goto)
}
\arguments{
\item{url}{Enedis URL for login
It defaults to \url{"https://espace-client-connexion.enedis.fr/auth/UI/Login"}
but other values may be useful for unit tests}

\item{realm}{passed as a query to enedis website. 
Only tested with "particuliers"}

\item{goto}{passed as a query to enedis website. 
Usually \url{"https://espace-client-particuliers.enedis.fr/group/espace-particuliers/accueil"}}
}
\value{
the result of httr::GET query
}
\description{
Access Enedis website and check we can reach it.
}
\details{
Accède au site d'Enedis et vérifie qu'on peut l'atteindre.
}
\examples{
access_enedis()
}
\seealso{
\code{\link{connect_enedis}}, \code{\link{disconnect_enedis}}
}
