% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_enedis.R
\name{connect_enedis}
\alias{connect_enedis}
\title{connect_enedis}
\usage{

  connect_enedis(enedis_url = paste("https://espace-client-connexion.enedis.fr",
  "auth/UI/Login", sep = "/"), realm = "particuliers",
  goto = paste("https://espace-client-particuliers.enedis.fr",
  "group/espace-particuliers/accueil", sep = "/"),
  secretfile = ".secret_enedis_json")
}
\arguments{
\item{enedis_url}{defaults to "https://espace-client-connexion.enedis.fr/auth/UI/Login"}

\item{realm}{defaults to "particuliers"}

\item{goto}{defaults to "https://espace-client-particuliers.enedis.fr/group/espace-particuliers/accueil"}

\item{secretfile}{path to your secretfile in json format. chemin d'accès à secretfile.}
}
\value{
the result of httr::GET or httr::POST query
}
\description{
Connection au site d'enedis. Cette fonction suppose au prealable d'avoir cree
un fichier "secretfile" au format json qui contient ses identifiants. Ce fichier
peut contenir quelque chose comme:
{
  "IDToken1": ["mon.mail@pour.enedis"],
  "IDToken2": ["mon.mot.de.passe.pour.enedis"]
}
}
\examples{
connect_enedis(secretfile = "~/.secret_enedis_json")
}
