% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_enedis.R
\name{disconnect_enedis}
\alias{disconnect_enedis}
\title{disconnect_enedis}
\usage{
disconnect_enedis(url, url_logout)
}
\arguments{
\item{url}{Enedis url used to log in}

\item{url_logout}{Enedis url to be fetched for logout}
}
\value{
the result of httr::GET query
}
\description{
Disconnect from enedis website
}
\examples{
connect_enedis()
disconnect_enedis()
}
\seealso{
\code{\link{connect_enedis}}
}
