% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enedis_accueil}
\alias{enedis_accueil}
\title{Enedis connected mode URL}
\format{A character string with the url
\describe{
  \item{protocol}{https}
  \item{site}{espace-client-particuliers.enedis.fr}
  \item{path}{/group/espace-particuliers/accueil}
}}
\source{
\url{https://espace-client-particuliers.enedis.fr/group/espace-particuliers/accueil}
}
\usage{
enedis_accueil
}
\description{
Uniform Ressource Locator for Enedis connected mode welcome screen
}
\keyword{datasets}
