% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enedis_url}
\alias{enedis_url}
\title{Enedis URL}
\format{A character string with the url
\describe{
  \item{protocol}{https}
  \item{site}{espace-client-connexion.enedis.fr}
  \item{path}{/auth/UI/Login}
}}
\source{
\url{https://espace-client-connexion.enedis.fr/auth/UI/Login}
}
\usage{
enedis_url
}
\description{
Uniform Ressource Locator for Enedis login dialog
}
\keyword{datasets}
