% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enedis_url_logout}
\alias{enedis_url_logout}
\title{Enedis disconnect URL}
\format{A character string with the url
\describe{
  \item{protocol}{https}
  \item{site}{espace-client-particuliers.enedis.fr}
  \item{path}{/c/portal/logout}
}}
\source{
\url{https://espace-client-particuliers.enedis.fr/c/portal/logout}
}
\usage{
enedis_url_logout
}
\description{
Uniform Ressource Locator for Enedis disconnection
}
\keyword{datasets}
