% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_enedis.R
\name{access_enedis}
\alias{access_enedis}
\title{access_enedis}
\usage{
access_enedis(enedis_url = paste("https://espace-client-connexion.enedis.fr",
  "auth/UI/Login", sep = "/"), realm = "particuliers",
  goto = paste("https://espace-client-particuliers.enedis.fr",
  "group/espace-particuliers/accueil", sep = "/"))
}
\arguments{
\item{enedis_url}{defaults to "https://espace-client-connexion.enedis.fr/auth/UI/Login"}

\item{realm}{defaults to "particuliers"}

\item{goto}{defaults to "https://espace-client-particuliers.enedis.fr/group/espace-particuliers/accueil"}
}
\value{
the result of httr::GET query
}
\description{
Access Enedis website and check we can reach it.
}
\details{
Accède au site d'Enedis et vérifie qu'on peut l'atteindre.
}
\examples{
access_enedis()
}
