% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_enedis.R
\name{change_dates}
\alias{change_dates}
\title{change_dates}
\usage{
change_dates(form_query, end_date)
}
\arguments{
\item{form_query}{a list with url, query and input like daily_month}

\item{end_date}{the end date for the 1-month query in Date format}
}
\value{
an updated form query
}
\description{
Alter the start and end date for a daily 1-month query
}
\examples{
change_dates()
}
\seealso{
\code{\link{daily_month}}
}
