% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_enedis.R
\name{connect_enedis}
\alias{connect_enedis}
\title{connect_enedis}
\usage{
connect_enedis(url, realm = "particuliers", goto,
  secretfile = "~/.secret_enedis_json")
}
\arguments{
\item{url}{defaults to "https://espace-client-connexion.enedis.fr/auth/UI/Login"}

\item{realm}{defaults to "particuliers"}

\item{goto}{defaults to "https://espace-client-particuliers.enedis.fr/group/espace-particuliers/accueil"}

\item{secretfile}{path to your secretfile in json format. chemin d'accès à secretfile.}
}
\value{
the result of httr::GET or httr::POST query
}
\description{
Connects to Enedis website. This function searches identifier and password in a "secretfile" in json format.
Connection au site d'enedis. Cette fonction suppose au prealable d'avoir cree
un fichier "secretfile" au format json qui contient ses identifiants.
}
\details{
The secretfile should look like:
Le fichier secretfile peut contenir quelque chose comme :

\preformatted{
\{   
  "IDToken1": ["mon.mail@connu.d.enedis"],   
  "IDToken2": ["mon.mot.de.passe.pour.enedis"]   
\}   
}
}
\examples{
connect_enedis(secretfile = "~/.secret_enedis_json")
disconnect_enedis()
}
\seealso{
\code{\link{disconnect_enedis}}

\code{\link{query_daily_month}}
}
